updating jenkins master/agent images
