--- resources/jenkins-master/README.md
@@ -1,1 +1,0 @@
-
