+++ resources/jenkins-master/resources/plugins/plugins.txt
@@ -1,125 +1,128 @@
 ace-editor:1.1
-analysis-core:1.96
+analysis-core:1.95
 ant:1.9
 antisamy-markup-formatter:1.5
 apache-httpcomponents-client-4-api:4.5.5-3.0
-artifactory:3.2.2
+artifactory:3.2.0
 authentication-tokens:1.3
-blueocean:1.16.0
-blueocean-autofavorite:1.2.4
-blueocean-bitbucket-pipeline:1.16.0
-blueocean-commons:1.16.0
-blueocean-config:1.16.0
-blueocean-core-js:1.16.0
-blueocean-dashboard:1.16.0
-blueocean-display-url:2.3.0
-blueocean-events:1.16.0
-blueocean-executor-info:1.16.0
-blueocean-git-pipeline:1.16.0
-blueocean-github-pipeline:1.16.0
-blueocean-i18n:1.16.0
-blueocean-jira:1.16.0
-blueocean-jwt:1.16.0
-blueocean-personalization:1.16.0
-blueocean-pipeline-api-impl:1.16.0
-blueocean-pipeline-editor:1.16.0
-blueocean-pipeline-scm-api:1.16.0
-blueocean-rest:1.16.0
-blueocean-rest-impl:1.16.0
-blueocean-web:1.16.0
+blueocean:1.11.1
+blueocean-autofavorite:1.2.3
+blueocean-bitbucket-pipeline:1.11.1
+blueocean-commons:1.11.1
+blueocean-config:1.11.1
+blueocean-core-js:1.11.1
+blueocean-dashboard:1.11.1
+blueocean-display-url:2.2.0
+blueocean-events:1.11.1
+blueocean-git-pipeline:1.11.1
+blueocean-github-pipeline:1.11.1
+blueocean-i18n:1.11.1
+blueocean-jira:1.11.1
+blueocean-jwt:1.11.1
+blueocean-personalization:1.11.1
+blueocean-pipeline-api-impl:1.11.1
+blueocean-pipeline-editor:1.11.1
+blueocean-pipeline-scm-api:1.11.1
+blueocean-rest:1.11.1
+blueocean-rest-impl:1.11.1
+blueocean-web:1.11.1
 bouncycastle-api:2.17
-branch-api:2.5.2
+branch-api:2.1.2
 build-timeout:1.19
-cloudbees-bitbucket-branch-source:2.4.4
-cloudbees-folder:6.8
+cloudbees-bitbucket-branch-source:2.4.2
+cloudbees-folder:6.7
 command-launcher:1.3
-config-file-provider:3.6
-credentials:2.1.19
-credentials-binding:1.18
+config-file-provider:3.5
+configuration-as-code:1.5
+configuration-as-code-support:1.5
+credentials:2.1.18
+credentials-binding:1.17
 dependency-check-jenkins-plugin:4.0.2
 dependency-track:2.1.0
-display-url-api:2.3.1
-docker-commons:1.15
-docker-workflow:1.18
+display-url-api:2.3.0
+docker-commons:1.13
+docker-workflow:1.17
 durable-task:1.29
-email-ext:2.66
+email-ext:2.63
 favorite:2.3.2
 filesystem_scm:2.1
-git:3.10.0
-git-client:2.7.7
+git:3.9.3
+git-client:2.7.6
 git-server:1.7
 github:1.29.4
 github-api:1.95
-github-branch-source:2.5.3
-gradle:1.32
+github-branch-source:2.4.2
+gradle:1.30
 handlebars:1.1.1
-handy-uri-templates-2-api:2.1.7-1.0
+handy-uri-templates-2-api:2.1.6-1.0
 htmlpublisher:1.18
-ivy:2.1
-jackson2-api:2.9.9
-javadoc:1.5
+ivy:1.28
+jackson2-api:2.9.8
+javadoc:1.4
 jdk-tool:1.2
-jenkins-design-language:1.16.0
-jira:3.0.7
-job-dsl:1.74
+jenkins-design-language:1.11.1
+jira:3.0.5
+job-dsl:1.71
 jquery:1.12.4-0
 jquery-detached:1.2.1
 jsch:0.1.55
-junit:1.28
+junit:1.27
+kubernetes:1.14.5
 kubernetes-credentials:0.4.0
 ldap:1.20
-lockable-resources:2.5
+lockable-resources:2.4
 mailer:1.23
 mapdb-api:1.0.9.0
-matrix-auth:2.4.2
-matrix-project:1.14
+matrix-auth:2.3
+matrix-project:1.13
 maven-plugin:3.2
-mercurial:2.6
+mercurial:2.5
 momentjs:1.1.1
-openshift-client:1.0.30
-openshift-login:1.0.17
-pam-auth:1.5.1
-pipeline-build-step:2.9
+openshift-login:1.0.16
+openshift-sync:1.0.31
+pam-auth:1.4
+pipeline-build-step:2.7
 pipeline-github-lib:1.0
-pipeline-graph-analysis:1.10
-pipeline-input-step:2.10
+pipeline-graph-analysis:1.9
+pipeline-input-step:2.9
 pipeline-milestone-step:1.3.1
-pipeline-model-api:1.3.8
+pipeline-model-api:1.3.4.1
 pipeline-model-declarative-agent:1.1.1
-pipeline-model-definition:1.3.8
-pipeline-model-extensions:1.3.8
-pipeline-rest-api:2.11
+pipeline-model-definition:1.3.4.1
+pipeline-model-extensions:1.3.4.1
+pipeline-rest-api:2.10
 pipeline-stage-step:2.3
-pipeline-stage-tags-metadata:1.3.8
-pipeline-stage-view:2.11
-pipeline-utility-steps:2.3.0
+pipeline-stage-tags-metadata:1.3.4.1
+pipeline-stage-view:2.10
+pipeline-utility-steps:2.2.0
 plain-credentials:1.5
 pubsub-light:1.12
 resource-disposer:0.12
-scm-api:2.4.1
-script-security:1.59
-sonar:2.9
+scm-api:2.3.0
+script-security:1.53
+sonar:2.8.1
 sse-gateway:1.17
-ssh-credentials:1.16
+ssh-credentials:1.14
 ssh-slaves:1.29.4
-structs:1.19
+structs:1.17
 subversion:2.12.1
-swarm:3.16
-templating-engine:1.1.1
+swarm:3.15
+templating-engine:1.0
 timestamper:1.9
-token-macro:2.7
+token-macro:2.6
+trilead-api:1.0.1
 variant:1.2
 windows-slaves:1.4
 workflow-aggregator:2.6
-workflow-api:2.34
-workflow-basic-steps:2.16
-workflow-cps:2.68
+workflow-api:2.33
+workflow-basic-steps:2.14
+workflow-cps:2.63
 workflow-cps-global-lib:2.13
-workflow-durable-task-step:2.30
-workflow-job:2.32
-workflow-multibranch:2.21
+workflow-durable-task-step:2.29
+workflow-job:2.31
+workflow-multibranch:2.20
 workflow-remote-loader:1.4
 workflow-scm-step:2.7
 workflow-step-api:2.19
-workflow-support:3.3
+workflow-support:3.2
 ws-cleanup:0.37
