--- resources/jenkins-master/resources/scripts/jenkins-preboot.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-# Copyright © 2018 Booz Allen Hamilton. All Rights Reserved.
-# This software package is licensed under the Booz Allen Public License. The license can be found in the License file or at http://boozallen.github.io/licenses/bapl
-
-# Install the JTE Plugin
-# cp /usr/share/jenkins/ref/jte.jpi $JENKINS_HOME/plugins
